requirements.txt
django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-pytest==0.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.8
pytest==7.4.2
pytest-cov==4.1.0
pytest-factoryboy==2.5.1
python-dotenv==1.0.0
